refactor: move cp read func to cp_info
